Add delivery_method when creating SummaryDocumentActivity
This should have a value of 'postal' or 'digital' depending
on the type of summary document that was generated.

Passing in an invalid `delivery_method` will immediately
raise an error, while passing in an invalid appointment_id
or owner_uid will not, instead the save method will fail
and an error value will being set.

This is due on one being a code error (delivery_method),
while the other can be caused by the state of
the data in an external system.

ie:

* User has never logged into source system
* Appointment is in an invalid state